renamed helpers test into own module
